Set up ESLint
Tools that help us avoid problems in our code and maintain stylistic
consistency are helpful. We have been using ESLint to good effect in the
Brigade project and some of our other open source projects, so it seems
reasonable to set up the same thing here.

Most of the configuration was copied from the react-waypoint project and
modified slightly to fit this project. I ended up disabling a couple of
rules that were particularly annoying for this project's style that we
might want to consider but didn't seem worth it at this time. These
were:

- no-shadow
- no-use-before-define
- valid-jsdoc

I ended up removing the old JSCS and JSHint configuration files since
this tool fills the same role as those tools.

I also added the .travis.yml file to the .npmignore, since people who
install the package won't need that file.